Change where steps are performed
Lab 2 says to perform steps on Jumpbox, this lab says to use Azure Cloud Shell, and the next lab says to use the Jumpbox. This results in the user not being connected to AKS when starting Lab 4. Connecting to AKS is part of Lab 3 which is currently done in Cloud Shell.